[Dashboard] Increase the RPC timeout for the snapshot API (#28330)
The current RPC timeout is too short (2s), and we've discovered Ray components not responding within the current timeout range occasionally under pressure. This is something we will fix, but it's better if we can have a longer timeout. The current timeout is configured as 2X of the polling frequency.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>