[air] Add `Checkpoint.uri` to return checkpoint URI, if available (#28731)
A common ask is to retrieve the URI of a cloud checkpoint, e.g. after training. This PR introduces a property to the `Checkpoint` class that will return a URI if available and reachable from the local node (i.e. cloud storage or locally available file).

If accepted, we should then return URI checkpoints from Tune if syncing to cloud is enabled.

Signed-off-by: Kai Fricke <kai@anyscale.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>