[AIR] Support large checkpoints and other arguments (#28826)
Signed-off-by: Amog Kamsetty amogkamsetty@yahoo.com

Previously the arguments passed to the Trainer would be captured in the Trainable context. For arguments that are very large in size, this would prevent the Trainable from being registered due to gRPC resource limits.

Instead, we now always use tune.with_parameters to save the Trainer arguments in the object store rather than capturing it in the context.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>