Progress bar (#28575)
Progress bar for showing number of tasks completed /  running / scheduled.
Creates dashboard api endpoint which proxies to prometheus to return data for the progress bar.
Requires prometheus to be running.

Light dependency on #28286 since that will help make prometheus easier to run out of the box.

Per job progress is shown under the "progress" column in the table of jobs.

<img width="1904" alt="Screen Shot 2022-09-28 at 6 16 15 PM" src="https://user-images.githubusercontent.com/711935/192916737-e9c8dde0-18e5-4ff8-b767-1091dea57287.png">
<img width="1904" alt="Screen Shot 2022-09-28 at 6 16 44 PM" src="https://user-images.githubusercontent.com/711935/192916757-c87e9eee-ec42-4729-a9b2-74044617122d.png">

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>