[ci/release] Nightly long running should be smoke tests, fix nightly-3x (#29042)
First, this fixes a bug where the `nightly-3x` string was parsed incorrectly, which effectively started nightly tests and not the nightly-3x tests.

Second, this activates smoke tests for the long running tests for nightly tests, and moves the full tests to a weekly schedule. This is to avoid occupying too many instances during the week and save costs. Also, most failures happen earlier in the run, so it is sufficient to test a 24h workload once a week.

Signed-off-by: Kai Fricke <kai@anyscale.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>