[core] Mark `run_function_on_all_workers` as deprecated until we get rid of this. (#29062)
run_function_on_all_workers has a lot of issues:

no guarantee that the function run before the remote function run.
pubsub signal might be lost in some failure cases.
We can't delete it for now because the working dir init depends on this. So mark it as deprecated to warn people who uses this and delete it once we fix the working dir.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>