[core] Fix bug in dynamic generator tasks in object location handling (#29082)
When dynamically generated objects are put into the object store, we need to notify the owner of the new location. However, currently this message can reach the owner before the owner knows about the object, since the task reply can arrive much later. Then, the owner will think that the object has already gone out of scope and will not add the location, leading to an ObjectFetchTimedOutError once any worker on a different node tries to read the object.

This PR fixes the bug by adding the object to the ref counter if it was returned by a dynamic generator task. The owner does this by checking whether the task spec had num_returns="dynamic" set.
Related issue number

Closes #28911.

Signed-off-by: Stephanie Wang <swang@cs.berkeley.edu>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>