[Train] Make `prepare_model` always use the correct device (#29104)
Signed-off-by: Amog Kamsetty amogkamsetty@yahoo.com

Previously, prepare_model would use the local rank as the device even though local rank may not be the same as the actual device index. This mismatch can happen when CUDA_VISIBLE_DEVICES is set for example, which we do by default in Ray Train.

We should always use train.torch.get_device() as the device values for wrapping in DDP.

Closes #28996

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>