[Datasets] Bundle blocks smaller than batch size in `map_batches` tasks. (#28648)
When blocks are smaller than the batch size in map_batches(), the actual size of batches provided to the UDF can be much smaller (order of magnitude) than the specified batch size, resulting in very poor batch mapping throughput when the UDF's performance is sensitive to the batch size (such as in batch inference on GPUs).

This PR optimistically sends multiple blocks to a single mapper task up to (but not past) the provided batch size, mitigating the case in which block size << batch size without needing a shuffle step.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>