[Train] Support full `ray.get_gpu_ids()` API in `train.torch.get_device()` (#28659)
Signed-off-by: Amog Kamsetty amogkamsetty@yahoo.com

ray.get_gpu_ids() sometimes returns a list of ints or sometimes returns a list of strings depending on if the user has set the CUDA_VISIBLE_DEVICES environment variable. This has led to the following issue: #28467.

It seems like solidifying the Ray Core API will require more discussion (follow the thread here: #28632), so we temporarily account for this in Ray Train for now.

Closes #28467

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>