[workflow] Fix the object loss due to driver exit issues. (#29092)
When the workflow runs in driver mode, the owner of the object ref is the driver. So when the driver exits, the objects are no longer available. This happens when we run with `run_async`.

This PR fixed this by passing the manager actor as the owner of the objects.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>