[Metrics] Clean up metrics from a dead workers. (#29115)
This PR does several things.

Clean up MetricsAgent API and add class-level unit tests.
Add WorkerId to metrics report to figure out if the request is from workers.
New logic to clean up metrics when the worker is dead.
The logic is as follow; We use the metrics report as a heartbeat and clean up all related metrics when the report doesn't come on time.

In order to clean up metrics, we should clean up view_data.tag_value_aggregation_data_map[tag_vals]. In order to properly clean this up, we should track worker ID -> corresponding tag vals that it owns. We can guarantee 1 unique tuple of tag vals cannot be owned by multiple workers because it always contains the worker id

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>