[core] Support retry on python exception in dynamic generator tasks (#29159)
Supports retry_exception in dynamic generator tasks. This requires separate handling because for generator tasks, the actual execution happens when we iterate over the outputs, not when we call the task "function".

Will follow up with support for static generator tasks (num_returns=constant value), as well as doing other task-related handling like task cancellation for generators, but this is just a patch to unblock Datasets (which uses retry_exception to handle transient failures).

Signed-off-by: Stephanie Wang <swang@cs.berkeley.edu>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>