[Kubernetes][Doc] State intent to remove legacy Ray Operator (#29178)
This PR documents the intent to remove the legacy Ray operator in Ray 2.2.0.

Here's the timeline of our posture:

In Ray 2.0.0, we documented KubeRay as the preferred deployment method and stated the intent to deprecate
the legacy operator in the future.
In the Ray 2.1.0 release, we state that it is still possible to use the operator but give an explicit end-of-life at Ray 2.2.0.
In the Ray 2.2.0, we remove the operator code and documentation.

Signed-off-by: Dmitri Gekhtman <dmitri.m.gekhtman@gmail.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>