[runtime env] Allow __MACOS folder and multiple top-level directories for remote URI zip (#29010)
Implements the changes proposed in #28859, please see the issue for full details.

Copied from the issue:

If there is a single top level directory and possibly a __MACOS folder, stick to the current behavior and use the top level directory in the zip file as the working directory (this is the current behavior, but also fixes the __MACOS folder problem)
If there are multiple files or folders in the .zip file, take the contents of the zip file itself as the working directory.
This behavior will be fully backwards compatible and supports a common format of .zip files (i.e. put all the files into a top level directory), yet it also allows users to put files directly into the .zip file and removes the footgun.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>