[core] Fix the redis issue with TLS/SSL in AWS Redis (#29244)
When connecting to secured redis sometimes config is not allowed (https://redis.io/docs/manual/security/) by default for security reasons by the vendor.

In Ray, Redis used to be used for pubsub, but not anymore. This PR just deletes this config setup so that it can be used with vendor's redis.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>