[AIR] Ensure that driver `DatasetContext` is propagated to the `Trainer` actor. (#29192)
DatasetContext allows setting Datasets config values like the target max block size. Normally these are propagated from the driver to all tasks in the Dataset.

However, when using the AIR trainer, context changes made by the driver don't get propagated to the training dataset preprocessor. I think this is likely because the BatchMapper doesn't save the context, so when we pass the BatchMapper to the worker that actually creates the final Dataset, the driver's context values are lost.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>