[air/ci] Bump two test suite sizes (#29357)
We recently lowered the test instance sizes across the board. Two test suites began failing after #29295 were merged (at least one of the suites is only run on branch builds, so we missed it).

I'll run the tests on a repro CI instance to confirm they work on the larger instances.

Signed-off-by: Kai Fricke <kai@anyscale.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>