[ci] Use terraform-managed secret ARNs (#29369)
We've recently added these secrets to be managed with terraform. In this PR, we update the ARNs in our CI to use these terraform-managed ARNs instead of the old manual ones.
Additionally, we've moved the Ray AIR secrets into a single secret. The main reason for this is that there is a limit of policies that can be attached to a role, and by consolidating these similar secrets that will always share the same scope, we make sure not to run out of "space" in the future.

Signed-off-by: Kai Fricke <kai@anyscale.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>