[nightly-test][tune] fix tune test failures due to protobuf version change. (#29440)
Why are these changes needed?
the tune related nightly tests began to fail after we loose the upper bound limit for protobuf. Reading the log it looks the tensorboardX we installed is not compatible with latest protobuf. We need to explicit install tensorboardX so pip can recognize the dependency constraints of tensorboardX

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>