[Core] Fix parallel schedule pg hangs (#29491)
Our PG scheduling compose of 3 steps.

(optimization) select the feasible node candidates that could fulfill the request requirement
(optimization) scoring the node candidates based on their availability
(correctness) send the request to the selected node to acquire the resource. if failed, go back to step 1.
However, there is a bug in the current code where at step 1, we pick the node based on their availability instead of feasibility, this resulted in some pg marked infeasible.

this PR is based on #29355 but fixed the test failures.

Signed-off-by: 稚鱼 <lianjunwen.ljw@antgroup.com>
Co-authored-by: 稚鱼 <lianjunwen.ljw@antgroup.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>