[AIR - Datasets] [Hotfix] Tensor extension column concatenation fixes. (#29479)
Fixes the concatenation of tensor columns (and extension types in general). Previously, concatenating tensor columns may produce a variable-shaped tensor but would be represented with a broken homogeneous-shaped tensor. This PR ensures that the correct tensor extension type is produced post-concatenation.

Issue Details:

The core issue is that we weren’t accounting for homogeneous-shaped tensor column concatenation resulting in a heterogeneous-shaped tensor column. In master, this currently fails silently until accessing the data fails somewhere downstream of the concatenation.

E.g., if you have 5 images of shape (32, 32) in one block, and 5 images of shape (64, 64) in another block, the image column in each individual block is homogeneous-shaped. But if you concatenate those blocks, you have 10 images, half with shape (32, 32) and half with shape (64, 64), which needs our heterogeneous-shaped (variable-shaped) tensor column representation.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>