[AIR] Init Mosaic Trainer API (#29237)
Signed-off-by: ilee300a [ilee300@anyscale.com](mailto:ilee300@anyscale.com)

In this PR, we provide initial commits for integrating Mosaic library with Ray. As Mosaic library provides algorithmic acceleration, providing further acceleration from the system side via Ray's distributed training, we can improve the speedup of training process.

Included in this PR is MosaicTrainer skeleton code. For this PR, the trainer does not support using ray dataset shards in the worker loop and assumes that the data loaders are prepared in the trainer init function. The current trainer is able to run a composer model with callbacks and loggers as well as select algorithms. No metrics or checkpoints are reported from the trainer at the moment.

Co-authored-by: Amog Kamsetty <amogkamsetty@yahoo.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>