[AIR] Fix wandb callback failing with RLLib (#29731)
The logic in `WandbLoggerCallback` responsible for cleaning the object passed to `wandb` wasn't foolproof - namely, it allowed sets which are not JSON seriallizable to sneak through when passed to `wandb.init`, causing a JSON encoding error in `wandb`. This PR changes our serializability check to use the actual wandb JSON encoder, automatically converts sets to lists and adds testing.

Signed-off-by: Antoni Baum <antoni.baum@protonmail.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>