[autoscaler][aws/gcp/azure][job submission] Fix autoscaler defaults (#29741)
This fixes the default autoscaler configs so that the minimal examples are compatible (or more compatible with job submission). In particular, there are 2 fixes

Install ray[default] extras because the job submission portion of the api server depends on it.
Set ray start --head --dashboard-host=0.0.0.0 so that the dashboard listens for remote connections (instead of the default localhost).

Co-authored-by: Alex <alex@anyscale.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>